Bug fixes to ChestX loader, includes changes to vgg19 to be compatible with 224 images

+++ to_train/loss/resnet_loss.py
@@ -105,7 +105,7 @@ class ExitEnsembleDistillation(_MultiExitAccuracy):
         avg_acc = 0
         for (x,y) in val_loader:
             x = x.to(device)
-            y = y.to(device)
+            y = y.to(device, dtype=torch.long)
             target = y
             with torch.no_grad():
                 _ = model(x)
